暂时解决一下 Node上挂 有已经删除 的脚本

+++ cocos2d/core/CCNode.js
@@ -515,7 +515,12 @@ var Node = cc.Class({
         for (i = 0, len = this._components.length; i < len; ++i) {
             var component = this._components[i];
             // destroy immediate so its _onPreDestroy can be called
-            component._destroyImmediate();
+            if (component._destroyImmediate){
+                component._destroyImmediate();
+            }else{
+                console.log("coocs error on CCNode ",this.name," comp ",component.name);
+            }
+            
         }
 
         // Actions
